Save the logo ID automatically when it's dirty
The `save` function returned from the hook was memoized based on the
value before the new one was set, meaning that if we called it after
setting a new value it would save the old one.

To fix this we check is the prop is dirty when the component is next
rendered and then save it.